chore: move function definitions into submodule
